src/test/java/com/ys200209/reservationsystem/domain/service/ReservationServiceTest.java
package com.ys200209.reservationsystem.domain.service;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.when;

import com.ys200209.reservationsystem.domain.category.CategoriesResponseDto;
import com.ys200209.reservationsystem.domain.repository.JdbcReservationRepository;
import com.ys200209.reservationsystem.domain.repository.JdbcReservationRepositoryTest;
import com.ys200209.reservationsystem.domain.repository.ReservationRepository;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

class ReservationServiceTest {
    private CategoriesResponseDto categories = JdbcReservationRepositoryTest.categories;

    @Mock
    private ReservationRepository repository;

    @InjectMocks
    private ReservationService service;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void testApiCategories() {
        // when
        when(repository.getCategories()).thenReturn(categories);

        CategoriesResponseDto actual = service.getCategories();

        // then
        assertThat(actual).isEqualTo(categories);
    }
}
